requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.22.0,<0.23.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
mysql-connector-python>=8.0.0,<9.0.0
